+++ docs/source/releases.rst
@@ -8,6 +8,16 @@ This page details the changelog containing every notable change of every release
 - The release with "Unreleased" in title indicates that the release is not yet released and is under development.
 - The releases with "Pre-release" in title or if the version ends with an identifier, It indicates that the release was a pre-release.
 
+
+Latest
+------
+
+Additions
+~~~~~~~~~
+
+- Added :func:`qord.event` decorator for registering listeners in a subclassed :class:`Client`.
+
+
 v0.4.0
 ------